rename : change core file name to cores
