+++ src/frontend/src/components/cores/Icons/withSvgIcon.jsx
